creating booklist item

+++ composeApp/src/commonMain/kotlin/org/kb/bookpedia/book/domain/Book.kt
@@ -8,7 +8,7 @@ data class Book(
     val description: String,
     val languages: List<String>,
     val firstPublishYear: String?,
-    val avgRating: Int?,
+    val avgRating: Double?,
     val ratingCount: Int?,
     val numPages: Int?,
     val numEditions: Int
